pyproject.toml
[project]
name = "db-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "mysql-connector-python>=9.3.0",
    "ollama>=0.5.1",
    "openai>=1.86.0",
    "pydantic>=2.11.5",
    "uvicorn>=0.34.3",
]
